+++ .github/workflows/www-docker-image copy.yml
@@ -1,0 +1,53 @@
+name: Docker Image CI for API
+
+on:
+  push:
+    branches: [ "main" ]
+    paths:
+    - 'www/**'
+  pull_request:
+    branches: [ "main" ]
+    paths:
+    - 'www/**'
+
+env:
+  IMAGE_NAME: kennethgrace/netgpt-webui
+
+jobs:
+  build-and-push:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@2b82ce82d56a2a04d2637cd93a637ae1b359c0a7
+        
+      - name: Docker Metadata action
+        id: meta
+        uses: docker/metadata-action@2c0bd771b40637d97bf205cbccdd294a32112176
+        with:
+          images: ${{ env.IMAGE_NAME }}
+          tags: |
+            latest
+            type=sha
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@6a58db7e0d21ca03e6c44877909e80e45217eed2
+
+      - name: Login to Docker Hub
+        uses: docker/login-action@f4ef78c080cd8ba55a85445d5b36e214a81df20a
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_HUB_PASSWORD }}
+
+      - name: Build and push the API
+        uses: docker/build-push-action@44ea916f6c540f9302d50c2b1e5a8dc071f15cdf
+        with:
+          context: ./www
+          push: true
+          platforms: linux/amd64,linux/arm64
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+
