+++ www/app/src/components/App.tsx
@@ -4,10 +4,20 @@ import React, {
   LazyExoticComponent,
   Suspense,
   useEffect,
+  useMemo,
   useState,
 } from "react";
 
-import { AppBar, Box, IconButton, Toolbar, Typography } from "@mui/material";
+import {
+  AppBar,
+  Box,
+  Divider,
+  IconButton,
+  Paper,
+  Stack,
+  Toolbar,
+  Typography,
+} from "@mui/material";
 
 import { GitHub } from "@mui/icons-material";
 
@@ -19,6 +29,7 @@ import {
   Aliases,
   LanguageSettings,
   DefaultConfiguration,
+  useConfiguration,
 } from "../common/configuration";
 
 import {
@@ -28,81 +39,31 @@ import {
 } from "../data/browserCache";
 
 import { useFadeIn } from "../styles/effects";
+import { getManifest, ManifestFile } from "../data/appManifest";
 
-import "../styles/effect.css";
+import "../styles/fading.css";
 
-const ChatView = lazy(() => import("./ChatView"));
-const DialogsMenu = lazy(() => {
-  return new Promise((resolve) => setTimeout(resolve, 2000)).then(
-    () => import("./DialogsMenu")
-  );
-});
-
-const ApplicationTitle: string = "NetGPT";
-const ApplicationGithub: string = "https://github.com/kennethgrace/netgpt";
+const ChatView = lazy(() => import("./Viewport"));
+const Banner = lazy(() => import("./Banner"));
 
 const MidScreenLogo = () => {
-  return <Box
-    sx={{
-      position: "fixed",
-      top: "50%",
-      left: "50%",
-      transform: "translate(-50%, -50%)",
-      opacity: 0.4,
-    }}
-  >
-    <img alt="logo" aria-label="logo" src="/logo256x256.png" />
-  </Box>
-}
-
-interface BannerProps {
-  setOpenConfigDialog: (dialog: LazyExoticComponent<FC<any>>) => void;
-}
-
-const Banner: FC<BannerProps> = ({ setOpenConfigDialog }) => {
-  const handleGitHub = () => {
-    window.open(ApplicationGithub, "_blank");
-  };
-
   return (
-    <AppBar
-      position="sticky"
+    <Box
       sx={{
-        backgroundColor: "primary.main",
-        flexDirection: "row",
-        padding: "8px",
-        height: "64px",
-        alignItems: "center",
+        position: "fixed",
+        top: "50%",
+        left: "50%",
+        transform: "translate(-50%, -50%)",
+        opacity: 0.4,
       }}
     >
-      <Toolbar
-        sx={{
-          justifyContent: "flex-start",
-        }}
-      >
-        <DialogsMenu setOpenConfigDialog={setOpenConfigDialog} />
-      </Toolbar>
-      <Toolbar
-        sx={{
-          flexGrow: 1,
-          justifyContent: "center",
-        }}
-      >
-        <Typography variant="h4" fontWeight="bold" noWrap>
-          {ApplicationTitle}
-        </Typography>
-      </Toolbar>
-      <Toolbar>
-        <IconButton size="medium" aria-label="GitHub" onClick={handleGitHub}>
-          <GitHub />
-        </IconButton>
-      </Toolbar>
-    </AppBar>
+      <img alt="logo" aria-label="logo" src="/logo256x256.png" />
+    </Box>
   );
 };
 
 export const App: FC = () => {
-  const fadeIn = useFadeIn();
+  const [fadeIn, setFadeIn] = useFadeIn();
   const [dialogOpen, setDialogOpen] = React.useState(false);
   const [OpenConfigDialog, setOpenConfigDialog] =
     React.useState<LazyExoticComponent<FC<any>>>();
@@ -114,6 +75,7 @@ export const App: FC = () => {
   const [networkSettings, setNetworkSettings] = useState<NetworkSettings>();
   const [languageSettings, setLanguageSettings] = useState<LanguageSettings>();
   const [aliases, setAliases] = useState<Aliases>(DefaultConfiguration.aliases);
+  const [manifest, setManifest] = useState<ManifestFile>();
 
   const configContext = {
     serverUrl,
@@ -126,7 +88,16 @@ export const App: FC = () => {
     setAliases,
   };
 
-  // Check the browser's local storage for a saved configuration
+  useEffect(() => {
+    getManifest().then((manifest) => {
+      if (manifest) {
+        setManifest(manifest);
+      }
+    });
+    setFadeIn(true);
+  }, []);
+
+  // Check the browser's local storage for a saved configuration on startup
   useEffect(() => {
     const serverUrl = loadServerUrl();
     const networkSettings = loadNetworkSettings();
@@ -139,31 +110,38 @@ export const App: FC = () => {
 
   return (
     <ConfigurationContext.Provider value={configContext}>
-      <Box className={fadeIn ? "fade-in" : ""}>
-        <MidScreenLogo />
-        <Box
-          component="main"
-          sx={{
-            display: "flex",
-            flexDirection: "column",
-            width: "100vw",
-            height: "100vh",
-          }}
-        >
-          <Banner
-            setOpenConfigDialog={(dialog) => {
-              setOpenConfigDialog(dialog);
-              setDialogOpen(true);
+      {manifest ? (
+        <Box className={fadeIn ? "fade-in" : ""}>
+          <MidScreenLogo />
+          <Box
+            component="main"
+            sx={{
+              display: "flex",
+              flexDirection: "column",
+              width: "100vw",
+              height: "100vh",
             }}
-          />
-          <Suspense fallback={<LoadingIndicatorWithBackdrop />}>
+          >
+            <Banner
+              title={manifest.short_name}
+              github_url={manifest.github_url}
+              license_url={manifest.license_url}
+              setOpenConfigDialog={(dialog) => {
+                setOpenConfigDialog(dialog);
+                setDialogOpen(true);
+              }}
+            />
+            <Suspense fallback={<LoadingIndicatorWithBackdrop />}>
             {OpenConfigDialog && (
               <OpenConfigDialog isOpen={dialogOpen} onClose={setDialogOpen} />
             )}
-          </Suspense>
-          <ChatView />
+            </Suspense>
+            <ChatView />
+          </Box>
         </Box>
-      </Box>
+      ) : (
+        <LoadingIndicatorWithBackdrop label="Waiting for Application Manifest" />
+      )}
     </ConfigurationContext.Provider>
   );
 };
