+++ www/app/src/json/defaultMessages.json
@@ -15,13 +15,10 @@
       "sections": [
         {
           "messageType": "text",
-          "content": "Hello, User!"
-        },
-        {
-          "messageType": "code",
-          "content": "To get started, please review your `Server Settings` in the upper left corner."
+          "content": "Hello, User!" 
         }
       ],
+      "caption": "To get started, please review your \"Server Settings\" in the upper left corner.",
       "timestamp": 200
     }
   ]
